Refactor Neo4j parser and loader and update to support array-based formats
In addition, this can now be used as a script, e.g:

python -m 'graphlib.neo4j' [path/to/file.json] [--load] [uri]

If no path is provided, stdin will be read. If the `--load` flag is present,
it will execute the statements in Neo4j at the default URI unless a URI
is passed. Otherwise statements will be printed to stdout.